Move to Scala 2.10.0-RC1
