+++ framework/project/Build.scala
@@ -333,7 +333,7 @@ object PlayBuild extends Build {
         val buildWithDoc      = Option(System.getProperty("generate.doc")).isDefined
         val previousVersion   = "2.0.3"
         val previousScalaVersion = "2.9.1"
-        val buildScalaVersion = "2.10.0-M7"
+        val buildScalaVersion = "2.10.0-RC1"
         val buildScalaVersionForSbt = "2.9.2"
         val buildSbtVersion   = "0.12.1"
         val buildSbtVersionBinaryCompatible = "0.12"
@@ -439,7 +439,7 @@ object PlayBuild extends Build {
             "javax.servlet"                     %    "javax.servlet-api"        %   "3.0.1",
             "javax.transaction"                 %    "jta"                      %   "1.1",
 
-            "org.specs2"                        %   "specs2_2.10.0-M7"          %   "1.12.1.1" %  "test"
+            "org.specs2"                        %   "specs2_2.10.0-RC1"         %   "1.12.2" %  "test"
         )
      
         val runtime = Seq(
@@ -452,14 +452,12 @@ object PlayBuild extends Build {
             "ch.qos.logback"                    %    "logback-core"             %   "1.0.7",
             "ch.qos.logback"                    %    "logback-classic"          %   "1.0.7",
             
-            "com.github.scala-incubator.io"     %%   "scala-io-file"            %   "0.4.1" exclude("javax.transaction", "jta"),
+            "com.github.scala-incubator.io"     %   "scala-io-file_2.10.0-RC1"  %   "0.4.1" exclude("javax.transaction", "jta"),
             
-            ("com.typesafe.akka"                 %    "akka-actor_2.10.0-M7"     %   "2.1-M2").exclude("com.typesafe.config", "config"),
-            ("com.typesafe.akka"                 %    "akka-slf4j_2.10.0-M7"     %   "2.1-M2").exclude("com.typesafe.config", "config"),
-            // Akka 2.1 RC will depend on config 1.0.0 so this won't be necessary
-            "com.typesafe"                      %    "config"                   %   "1.0.0",
+            "com.typesafe.akka"                 %    "akka-actor_2.10.0-RC1"    %   "2.1.0-RC1",
+            "com.typesafe.akka"                 %    "akka-slf4j_2.10.0-RC1"    %   "2.1.0-RC1",
 
-            "org.scala-tools"                   %    "scala-stm_2.10.0-M7"      %   "0.6",
+            "org.scala-stm"                     %    "scala-stm_2.10.0-RC1"     %   "0.6",
 
             "joda-time"                         %    "joda-time"                %   "2.1",
             "org.joda"                          %    "joda-convert"             %   "1.2",
@@ -480,7 +478,7 @@ object PlayBuild extends Build {
 
             "javax.transaction"                 %    "jta"                      %   "1.1",
 
-            "org.specs2"                        %   "specs2_2.10.0-M7"          %   "1.12.1.1" %  "test",
+            "org.specs2"                        %   "specs2_2.10.0-RC1"         %   "1.12.2" %  "test",
 
             "org.mockito"                       %    "mockito-all"              %   "1.9.0"    %  "test",
             "com.novocode"                      %    "junit-interface"          %   "0.8"      %  "test",
@@ -499,7 +497,7 @@ object PlayBuild extends Build {
 
         val templatesCompilerDependencies = Seq(
             "com.github.scala-incubator.io"     %%   "scala-io-file"            %   "0.4.1" exclude("javax.transaction", "jta"),
-            "org.specs2"                        %%   "specs2"                   %   "1.12.1"    %   "test"
+            "org.specs2"                        %%   "specs2"                   %   "1.12.2"    %   "test"
               exclude("javax.transaction", "jta")
         )
         
@@ -535,23 +533,23 @@ object PlayBuild extends Build {
         )
 
         val consoleDependencies = Seq(
-            "net.databinder.giter8"             %   "giter8_2.9.1"             %   "0.5.0" 
+            "net.databinder.giter8"             %   "giter8_2.9.1"              %   "0.5.0" 
         )
 
         val templatesDependencies = Seq(
-            "com.github.scala-incubator.io"     %%   "scala-io-file"            %   "0.4.1" exclude("javax.transaction", "jta"),
-            "org.specs2"                        %    "specs2_2.10.0-M7"         %   "1.12.1.1"    %   "test"
+            "com.github.scala-incubator.io"     %    "scala-io-file_2.10.0-RC1" %   "0.4.1" exclude("javax.transaction", "jta"),
+            "org.specs2"                        %    "specs2_2.10.0-RC1"        %   "1.12.2"    %   "test"
         )
 
         val iterateesDependencies = Seq(
-            "org.scala-tools"               %    "scala-stm_2.10.0-M7"      %   "0.6",
-            "com.github.scala-incubator.io"     %%   "scala-io-file"            %   "0.4.1" exclude("javax.transaction", "jta")
+            "org.scala-stm"                     %    "scala-stm_2.10.0-RC1"     %   "0.6",
+            "com.github.scala-incubator.io"     %    "scala-io-file_2.10.0-RC1" %   "0.4.1" exclude("javax.transaction", "jta")
         )
           
 
         val testDependencies = Seq(
             "junit"                             %    "junit-dep"                %   "4.10",
-            "org.specs2"                        %    "specs2_2.10.0-M7"         %   "1.12.1.1",
+            "org.specs2"                        %    "specs2_2.10.0-RC1"        %   "1.12.2",
             "com.novocode"                      %    "junit-interface"          %   "0.8" exclude ("junit", "junit"),
 
             "org.fluentlenium"                  %    "fluentlenium-festassert"  %   "0.7.3"
