+++ framework/src/play/src/main/scala/play/core/server/netty/NettyPromise.scala
@@ -3,7 +3,7 @@ package play.core.server.netty
 import org.jboss.netty.channel.ChannelFuture
 import play.api.libs.concurrent._
 import scala.concurrent.{ ExecutionContext, CanAwait }
-import scala.concurrent.util.Duration
+import scala.concurrent.duration.Duration
 import java.util.concurrent.TimeUnit
 import org.jboss.netty.channel.ChannelFutureListener
 import scala.util._
