+++ samples/java/comet-clock/app/controllers/Application.java
@@ -10,7 +10,7 @@ import akka.actor.*;
 
 import java.util.*;
 import java.text.*;
-import scala.concurrent.util.Duration;
+import scala.concurrent.duration.Duration;
 
 import static java.util.concurrent.TimeUnit.*;
