+++ samples/scala/comet-clock/app/controllers/Application.scala
@@ -7,7 +7,7 @@ import play.api.libs.{ Comet }
 import play.api.libs.iteratee._
 import play.api.libs.concurrent._
 
-import scala.concurrent.util.duration._
+import scala.concurrent.duration._
 
 object Application extends Controller {
