+++ samples/scala/websocket-chat/app/controllers/Application.scala
@@ -9,7 +9,7 @@ import play.api.libs.iteratee._
 import models._
 
 import akka.actor._
-import scala.concurrent.util.duration._
+import scala.concurrent.duration._
 
 object Application extends Controller {
