+++ samples/scala/websocket-chat/app/models/ChatRoom.scala
@@ -1,7 +1,7 @@
 package models
 
 import akka.actor._
-import scala.concurrent.util.duration._
+import scala.concurrent.duration._
 
 import play.api._
 import play.api.libs.json._
